Avoid android code on other platforms
